num: using template to implement swap for any integer

+++ src/net/net_endpoint.cpp
@@ -169,7 +169,7 @@ chen::net::endpoint& chen::net::endpoint::operator=(const struct ::sockaddr *ep)
         case AF_INET:
         {
             auto in = (struct ::sockaddr_in*)ep;
-            this->_addr = num::swap(static_cast<std::uint32_t>(in->sin_addr.s_addr));
+            this->_addr = num::swap(in->sin_addr.s_addr);
             this->_port = num::swap(in->sin_port);
         }
             break;
